solver,meta: refactor const-related code
Move const defs and functions to a separate files.

Get rid of -1 multiplier hack to handle unary minus.
